[Gimbalrotor][new] add new package for gimbalrotor robot based on trirotor package
