+++ robots/gimbalrotor/config/Battery.yaml
