+++ robots/gimbalrotor/config/MotorInfo.yaml
