+++ robots/gimbalrotor/config/NavigationConfig.yaml
