+++ robots/gimbalrotor/config/sensors/imu/spinal.yaml
