+++ robots/gimbalrotor/plugins/control_plugins.xml
@@ -1,0 +1,5 @@
+<library path="lib/libgimbalrotor_controller">
+  <class name="aerial_robot_control/gimbalrotor_controller" type="aerial_robot_control::GimbalrotorController" base_class_type="aerial_robot_control::ControlBase">
+    <description>plugin for gimbalrotor control </description>
+  </class>
+</library>
