+++ robots/gimbalrotor/plugins/navigation_plugin.xml
@@ -1,0 +1,5 @@
+<library path="lib/libgimbalrotor_navigation">
+  <class name="aerial_robot_navigation/gimbalrotor_navigation" type="aerial_robot_navigation::GimbalrotorNavigator" base_class_type="aerial_robot_navigation::BaseNavigator">
+    <description>plugin for gimbalrotor navigation </description>
+  </class>
+</library>
