+++ robots/gimbalrotor/robots/gimbalrotor.gazebo.xacro
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<robot
+    xmlns:xacro="http://www.ros.org/wiki/xacro" name="gimbalrotor" >
+  <xacro:arg name="robot_name" default="gimbalrotor" />
+
+  <xacro:include filename="$(find gimbalrotor)/robots/gimbalrotor.urdf.xacro" />
+
+  <!-- gazebo plugin for default controller and sensors -->
+  <xacro:include filename="$(find aerial_robot_simulation)/xacro/spinal.gazebo.xacro" />
+  <xacro:gazebo_spinal robot_name="$(arg robot_name)" mag_frame="fc" />
+
+</robot>
