+++ robots/gimbalrotor/urdf/mesh/kxr_flight_trirotor_base_link.dae
