+++ robots/gimbalrotor/urdf/mesh/kxr_flight_unit_base_link.dae
