+++ robots/gimbalrotor/urdf/mesh/thrust_link_left_straight.STL
