+++ robots/gimbalrotor/urdf/mesh/thrust_link_right_straight.STL
