--- robots/trirotor/config/RobotModel.yaml
@@ -1,5 +1,0 @@
-# robot model plugin
-robot_model_plugin_name: trirotor_robot_model # multirotor_robot_model
-
-wrench_mat_det_thre: 0
-
