--- robots/trirotor/config/Servo.yaml
@@ -1,32 +1,0 @@
-servo_controller:
-  gimbals:
-    ctrl_pub_topic: servo/target_states
-    torque_pub_topic: servo/torque_enable
-    angle_sgn: -1
-    angle_scale: 0.001534
-    zero_point_offset: 2047
-
-    controller1:
-      id: 1
-      name: gimbal1
-      simulation:
-        pid: {p: 5.0, i: 0.1, d: 0.1, i_clamp_max: 2.0, i_clamp_min: -2.0}
-        init_value: 0
-        type: effort_controllers/JointPositionController
-
-    controller2:
-      id: 2
-      name: gimbal2
-      simulation:
-        pid: {p: 5.0, i: 0.1, d: 0.1, i_clamp_max: 2.0, i_clamp_min: -2.0}
-        init_value: 0 
-        type: effort_controllers/JointPositionController
-
-    controller3:
-      id: 3
-      name: gimbal3
-      simulation:
-        pid: {p: 5.0, i: 0.1, d: 0.1, i_clamp_max: 2.0, i_clamp_min: -2.0}
-        init_value: 0 
-        type: effort_controllers/JointPositionController
-
