--- robots/trirotor/config/TrirotorControl_sim.yaml
@@ -1,48 +1,0 @@
-aerial_robot_control_name: aerial_robot_control/trirotor_controller
-
-controller:
-  torque_allocation_matrix_inv_pub_interval: 0.05
-  wrench_allocation_matrix_pub_interval: 0.1
-
-  xy:
-    p_gain: 2.0
-    i_gain: 0.05
-    d_gain: 1.5
-    limit_sum: 4.0
-    limit_p: 12.0
-    limit_i: 12.0
-    limit_d: 12.0
-
-  z:
-    p_gain: 5.0
-    i_gain: 1.0
-    d_gain: 2.5
-    limit_err_p: 1.0
-    limit_sum: 25.0 # N for clamping thrust force
-    limit_p: 25.0
-    limit_i: 25.0
-    limit_d: 25.0
-    force_landing_descending_rate: -0.5
-
-  roll_pitch:
-    p_gain: 15.0
-    i_gain: 1.0
-    d_gain: 10.0
-    limit_sum: 20.0
-    limit_p: 20.0
-    limit_i: 20.0
-    limit_d: 20.0
-
-    start_rp_integration_height: 0.01
-
-  yaw:
-    p_gain: 5.0
-    i_gain: 1.0
-    d_gain: 4.0
-    limit_err_p: 0.4
-    limit_sum: 20.0
-    limit_p: 20.0
-    limit_i: 20.0
-    limit_d: 20.0
-
-    need_d_control: true
