--- robots/trirotor/package.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>trirotor</name>
-  <version>1.3.2</version>
-  <description>The trirotor package</description>
-
-  <maintainer email="chou@jsk.imi.i.u-tokyo.ac.jp">Bakui Chou</maintainer>
-  <license>BSD</license>
-  <url type="website">https://github.com/jsk-ros-pkg/jsk_aerial_robot</url>
-  <author email="sugihara@jsk.imi.i.u-tokyo.ac.jp">Kazuki Sugihara</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>aerial_robot_control</build_depend>
-  <build_depend>aerial_robot_estimation</build_depend>
-  <build_depend>aerial_robot_model</build_depend>
-  <build_depend>aerial_robot_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>pluginlib</build_depend>
-
-  <run_depend>aerial_robot_control</run_depend>
-  <run_depend>aerial_robot_estimation</run_depend>
-  <run_depend>aerial_robot_model</run_depend>
-  <run_depend>aerial_robot_msgs</run_depend>
-  <run_depend>aerial_robot_simulation</run_depend>
-  <run_depend>aerial_robot_base</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>pluginlib</run_depend>
-
-  <export>
-    <aerial_robot_model plugin="${prefix}/plugins/robot_model_plugins.xml" />
-    <aerial_robot_control plugin="${prefix}/plugins/control_plugins.xml" />
-    <aerial_robot_control plugin="${prefix}/plugins/navigation_plugin.xml" />
-  </export>
-
-</package>
