--- robots/trirotor/plugins/control_plugins.xml
@@ -1,5 +1,0 @@
-<library path="lib/libtrirotor_controller">
-  <class name="aerial_robot_control/trirotor_controller" type="aerial_robot_control::TrirotorController" base_class_type="aerial_robot_control::ControlBase">
-    <description>plugin for trirotor control </description>
-  </class>
-</library>
