--- robots/trirotor/robots/trirotor.urdf.xacro
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<robot
-    xmlns:xacro="http://www.ros.org/wiki/xacro" name="trirotor" >
-
-  <xacro:include filename="$(find trirotor)/urdf/trirotor.urdf.xacro" />
-  <xacro:include filename="$(find trirotor)/urdf/common.xacro" />
-
-  <xacro:trirotor_center_link child1="1" child2="2" child3="3"/>
-  <xacro:trirotor_thrust_link self="1" rotor_direction="1"/>
-  <xacro:trirotor_thrust_link self="2" rotor_direction="-1"/>
-  <xacro:trirotor_thrust_link self="3" rotor_direction="1"/>
-
-</robot>
